Code clean up.

+++ dlls/util.cpp
@@ -1975,7 +1975,7 @@ void CSave :: WriteFunction( const char *pname, void **data, int count )
 {
 	const char *functionName;
 
-	functionName = NAME_FOR_FUNCTION( (uint32)*data );
+	functionName = NAME_FOR_FUNCTION( (uint32)(size_t)*data );
 	if ( functionName )
 		BufferField( pname, strlen(functionName) + 1, functionName );
 	else
